requirements.txt
aniso8601==8.0.0
click==7.1.1
Flask==1.1.1
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.1
Markdown==3.2.1
MarkupSafe==1.1.1
pytz==2019.3
six==1.14.0
Werkzeug==1.0.0
flask_swagger_ui
Adafruit-Blinka==4.8.1
adafruit-circuitpython-neopixel==6.0.0
adafruit-circuitpython-pypixelbuf==2.0.0
Adafruit-PlatformDetect==2.11.1
Adafruit-PureIO==1.1.5
python-daemon